Fix duplicate includes rule
