+++ src/test/java/liquibase/parser/ext/CustomXMLChangeLogSAXParserTest.java
@@ -1,16 +1,41 @@
 package liquibase.parser.ext;
 
 import com.wcg.liquibase.config.Config;
+import com.wcg.liquibase.config.rules.RuleRunner;
 import com.wcg.liquibase.resolvers.DefaultConfigParameterResolver;
+import com.wcg.liquibase.resolvers.RuleRunnerParameterResolver;
 import liquibase.exception.ChangeLogParseException;
+import liquibase.parser.core.ParsedNode;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
-@ExtendWith(DefaultConfigParameterResolver.class)
+@ExtendWith({DefaultConfigParameterResolver.class, RuleRunnerParameterResolver.class})
 class CustomXMLChangeLogSAXParserTest {
+
+    @Test
+    void shouldAllowTokenSchemaName(RuleRunner ruleRunner) {
+        CustomXMLChangeLogSAXParser customXMLChangeLogSAXParser = new CustomXMLChangeLogSAXParser();
+        ParsedNode parsedNode = mockParsedNode("${schema_name}");
+        try {
+            customXMLChangeLogSAXParser.checkSchemaName(parsedNode, ruleRunner);
+        } catch (ChangeLogParseException e) {
+            fail(e);
+        }
+    }
+
+    @Test
+    void shouldNotAllowRawSchemaName(RuleRunner ruleRunner) {
+        CustomXMLChangeLogSAXParser customXMLChangeLogSAXParser = new CustomXMLChangeLogSAXParser();
+        ParsedNode parsedNode = mockParsedNode("SCHEMA_NAME");
+        ChangeLogParseException changeLogParseException =
+                assertThrows(ChangeLogParseException.class, () -> customXMLChangeLogSAXParser.checkSchemaName(parsedNode, ruleRunner));
+
+        assertTrue(changeLogParseException.getMessage().contains("Must use schema name token, not SCHEMA_NAME"));
+    }
 
     @Test
     void shouldPreventDuplicateIncludes(Config config) throws Exception {
@@ -24,5 +49,12 @@ class CustomXMLChangeLogSAXParserTest {
         ChangeLogParseException changeLogParseException =
                 assertThrows(ChangeLogParseException.class, () -> parser.checkDuplicateIncludes("foo/bar/baz.xml", config));
         assertEquals("Changelog file 'foo/bar/baz.xml' was included more than once", changeLogParseException.getMessage());
+    }
+
+    private ParsedNode mockParsedNode(String value) {
+        ParsedNode parsedNode = mock(ParsedNode.class);
+        when(parsedNode.getName()).thenReturn("schemaName");
+        when(parsedNode.getValue()).thenReturn(value);
+        return parsedNode;
     }
 }
