+++ src/test/resources/integration/duplicate-includes-change.xml
@@ -1,0 +1,12 @@
+<databaseChangeLog
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.3.xsd">
+
+    <changeSet id="201808281509" author="luke.rogers">
+        <sql>
+            SELECT 1 FROM DUAL
+        </sql>
+    </changeSet>
+
+</databaseChangeLog>
