+++ src/test/resources/integration/duplicate-includes-nested-change.xml
@@ -1,0 +1,14 @@
+<databaseChangeLog
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.3.xsd">
+
+    <include relativeToChangelogFile="true" file="duplicate-includes-change.xml"/>
+
+    <changeSet id="201808102132" author="luke.rogers">
+        <sql>
+            SELECT 1 FROM DUAL
+        </sql>
+    </changeSet>
+
+</databaseChangeLog>
